episode-004-tools_for_time/calvin/requirements.txt
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.47.0
anthropic
rich